+++ app/assets/javascripts/filterrific/filterrific-jquery.js
@@ -19,6 +19,9 @@ if (typeof Filterrific === 'undefined') {
 
 // Define function to submit Filterrific filter form
 Filterrific.submitFilterForm = function(){
+  if (this.lastRequest) {
+    this.lastRequest.abort();
+  }
   var form = $(this).parents("form"),
       url = form.attr("action");
   // send before event
@@ -26,7 +29,7 @@ Filterrific.submitFilterForm = function(){
   // turn on spinner
   $('.filterrific_spinner').show();
   // Submit ajax request
-  $.ajax({
+  this.lastRequest = $.ajax({
     url: url,
     data: form.serialize(),
     type: 'GET',
@@ -76,7 +79,7 @@ Filterrific.submitFilterForm = function(){
       check();
       var ti = setInterval(check, frequency); // invoke check periodically
       // reset counter after user interaction
-      $this.bind('keyup click mousemove', reset); //mousemove is for selects
+      $this.bind('keyup keydown click mousemove', reset); //mousemove is for selects
     });
   };
 })(jQuery);
@@ -97,16 +100,3 @@ Filterrific.init = function() {
     Filterrific.submitFilterForm
   );
 };
-
-
-// Initialize event observers on document ready and turbolinks page:load
-jQuery(document).on('turbolinks:load', function() {
-  // Prevent double initilisation. With turbolinks 5 this function
-  // will be called twice: on 'ready' and 'turbolinks:load'
-  jQuery(document).off('ready page:load')
-  Filterrific.init();
-});
-
-jQuery(document).on('ready page:load', function() {
-  Filterrific.init();
-});