+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Filterrific major versions match the Ruby on Rails major versions they work with.
 
+## [5.3.0] - Feb. 02, 2021
+
+* Reset timer on keydown to avoid a request triggering immediately after you start typing.
+* Abort last AJAX request before firing a new one.
+* Don't initialize automatically. User can call Filterrific.init themselves to avoid double initialization with Turbolinks.
+
 ## [5.2.1] - Aug. 5, 2019
 
 * Fixed issue where uncommitted code was pushed to rubygems and broke array filters.